+++ ChangeLog.md
@@ -1,5 +1,5 @@
 2022-11-02 Conzxy
-* 基于`Bresenham's line drawing algorithm`实现画线函数
+* 基于`Bresenham's line drawing algorithm`实现`DrawLine()`
 
 2022-11-03 Conzxy
 * `Vec`, `Matrix`
@@ -7,3 +7,4 @@
 2022-11-04 Conzxy
 * 解析`*.obj`格式文件的`Model`
 * wireframe render
+* 基于顶点的`重心坐标(Barycentric coordinate)`实现`DrawTriangle()`